# Conflicts:
#	.idea/workspace.xml
#	src/com/launcher/LibraryController.java

+++ src/database/DatabaseHandler.java
@@ -82,32 +82,6 @@ public class DatabaseHandler {
         return null;
     }
 
-    public void addGame(Game game) {
-        connect();
-        try {
-            String gameImage = LibraryController.getImagePath(game.getGameImage());
-            String gameCoverImage = LibraryController.getImagePath(game.getGameCoverImage());
-
-            String query = "INSERT INTO games (name, description, genre, image, coverimage, exe, folder) VALUES (?, ?, ?, ?, ?, ?, ?)";
-
-            PreparedStatement pstmt = conn.prepareStatement(query);
-            pstmt.setString(1, game.getGameName());
-            pstmt.setString(2, game.getGameDescription());
-            pstmt.setString(3, game.getGameGenre());
-            pstmt.setString(4, gameImage);
-            pstmt.setString(5, gameCoverImage);
-            pstmt.setString(6, game.getExeLocation());
-            pstmt.setString(7, game.getFolderLocation());
-
-            pstmt.executeUpdate();
-
-        } catch (SQLException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
     public void updateGame(Game game) {
         String SQL = "UPDATE games SET description = ?, genre = ?, image = ?, coverimage = ?, exe = ?, folder = ? WHERE name = ?";
         connect();
@@ -126,13 +100,14 @@ public class DatabaseHandler {
         }
     }
 
-    public List<Game> getGames() {
+    public List<Game> getGames(String userName) {
         List<Game> games = new ArrayList<>();
-        String SQL = "SELECT * FROM games";
+        String SQL = "SELECT games.* FROM games JOIN user_games ON games.id = user_games.game_id JOIN users ON user_games.user_id = users.id WHERE users.userName = ?";
 
         connect();
-        try (Statement stmt = conn.createStatement();
-             ResultSet rs = stmt.executeQuery(SQL)) {
+        try (PreparedStatement pstmt = conn.prepareStatement(SQL)) {
+            pstmt.setString(1, userName);
+            ResultSet rs = pstmt.executeQuery();
 
             while (rs.next()) {
                 Game game = new Game();
@@ -152,11 +127,43 @@ public class DatabaseHandler {
         return games;
     }
 
-    public void deleteGame(String gameName) {
-        String SQL = "DELETE FROM games WHERE name = ?";
+    public void addGame(Game game, String userName) {
+        connect();
+        try {
+            String gameImage = LibraryController.getImagePath(game.getGameImage());
+            String gameCoverImage = LibraryController.getImagePath(game.getGameCoverImage());
+
+            String query = "INSERT INTO games (name, description, genre, image, coverimage, exe, folder) VALUES (?, ?, ?, ?, ?, ?, ?)";
+
+            PreparedStatement pstmt = conn.prepareStatement(query);
+            pstmt.setString(1, game.getGameName());
+            pstmt.setString(2, game.getGameDescription());
+            pstmt.setString(3, game.getGameGenre());
+            pstmt.setString(4, gameImage);
+            pstmt.setString(5, gameCoverImage);
+            pstmt.setString(6, game.getExeLocation());
+            pstmt.setString(7, game.getFolderLocation());
+            pstmt.executeUpdate();
+
+            String query2 = "INSERT INTO user_games (user_id, game_id) SELECT users.id, games.id FROM users, games WHERE users.userName = ? AND games.name = ?";
+            PreparedStatement pstmt2 = conn.prepareStatement(query2);
+            pstmt2.setString(1, userName);
+            pstmt2.setString(2, game.getGameName());
+            pstmt2.executeUpdate();
+
+        } catch (SQLException e) {
+            e.printStackTrace();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public void deleteGame(String gameName, String userName) {
+        String SQL = "DELETE FROM user_games WHERE game_id = (SELECT id FROM games WHERE name = ?) AND user_id = (SELECT id FROM users WHERE userName = ?)";
         connect();
         try (PreparedStatement pstmt = conn.prepareStatement(SQL)) {
             pstmt.setString(1, gameName);
+            pstmt.setString(2, userName);
             pstmt.executeUpdate();
         } catch (SQLException e) {
             System.out.println(e.getMessage());
